Level + standardCases. FileLogger and FileLoggerManager merged together. ID identifier removed from loggers for now. Public functions related to FileLogger removed from LoggerManager.
