+++ Sources/Logger/Loggers/ConsoleLogger.swift
@@ -9,12 +9,7 @@ import Foundation
 
 /// Pre-built logger that logs to the console.
 public class ConsoleLogger: Logging {
-    public let id: UUID
     public var levels: [Level] = [.info]
-
-    public init(id: UUID = UUID()) {
-      self.id = id
-    }
 
     public func log(_ logEntry: LogEntry) {
         print("\(logEntry)")
